feedback method implementation changed
